+++ packages/react-components/src/components/dropdown/index.ts
@@ -1,3 +1,0 @@
-export { Dropdown } from './dropdown'
-export { Dropdown as Popover } from './dropdown'
-export type { DropdownProps } from './dropdown'